Refactor public page metadata and SEO handling
- Updated home and pricing routes to use new public page metadata structure.
- Enhanced legal route to handle not found pages and return appropriate metadata.
- Introduced new SDK references documentation with detailed element descriptions.
- Added public page definitions for home, pricing, documentation, and legal pages.
- Implemented a new prerender module to consolidate public page paths.
- Adjusted TypeScript configurations to include new types and improve module detection.
