+++ .gitignore
@@ -177,8 +177,6 @@ script.bat
 /web/*.tsbuildinfo
 /web/node_modules/.tmp/
 /web/node_modules/
-/web/dist/
-/web/dist-ssr/
 /web/*.local
 /web/.vscode/*
 !/web/.vscode/extensions.json
