+++ web/package.json
@@ -37,7 +37,7 @@
         "react": "^19.2.7",
         "react-dom": "^19.2.7",
         "react-hook-form": "^7.81.0",
-        "react-router": "^8.2.0",
+        "react-router": "8.2.0",
         "tailwindcss": "^4.3.2",
         "ufo": "^1.6.4",
         "valtio": "^2.3.2",
