--- web/src/platform/docs/routes.tsx
@@ -1,18 +1,0 @@
-import { useLocation } from 'react-router';
-import NotFound from '@/platform/NotFound';
-import DocsLayout from '@/platform/docs/layout';
-import { DOC_PAGES } from '@/platform/docs/catalog';
-
-/** Resolves documentation paths inside the lazy-loaded documentation route group. */
-export default function DocumentationRoutes() {
-    const location = useLocation();
-
-    // Match the current path to its complete article definition.
-    const pathname = location.pathname.replace(/\/+$/, '') || '/';
-    const page = DOC_PAGES.find((item) => item.path === pathname);
-    if (!page) {
-        return <NotFound />;
-    }
-
-    return <DocsLayout page={page} />;
-}
