+++ web/src/platform/legal/catalog.tsx
@@ -1,10 +1,11 @@
 import { FileText, Landmark, ShieldCheck } from 'lucide-react';
 import type { ArticleNavigationGroup, ArticlePage } from '@/platform/catalog';
+import { homePage, legalPages } from '@/platform/public';
 import { content as termsContent, metadata as termsMetadata } from '@/platform/legal/terms';
 import { content as privacyContent, metadata as privacyMetadata } from '@/platform/legal/privacy';
 import { content as impressumContent, metadata as impressumMetadata } from '@/platform/legal/impressum';
 
-const homeBreadcrumb = { title: 'Home', path: '/' };
+const homeBreadcrumb = { title: 'Home', path: homePage.path };
 
 /** Builds a legal page with its standard Home breadcrumb. */
 function legalPage(page: Omit<ArticlePage, 'breadcrumbs'>): ArticlePage {
@@ -15,24 +16,21 @@ function legalPage(page: Omit<ArticlePage, 'breadcrumbs'>): ArticlePage {
 }
 
 const termsPage = legalPage({
-    title: 'Terms of Service',
-    path: '/terms',
+    ...legalPages.terms,
     icon: <FileText aria-hidden="true" size={16} />,
     content: termsContent,
     metadata: termsMetadata,
 });
 
 const impressumPage = legalPage({
-    title: 'Impressum',
-    path: '/impressum',
+    ...legalPages.impressum,
     icon: <Landmark aria-hidden="true" size={16} />,
     content: impressumContent,
     metadata: impressumMetadata,
 });
 
 const privacyPage = legalPage({
-    title: 'Privacy',
-    path: '/privacy',
+    ...legalPages.privacy,
     icon: <ShieldCheck aria-hidden="true" size={16} />,
     content: privacyContent,
     metadata: privacyMetadata,
