--- web/src/platform/paths.ts
@@ -1,67 +1,0 @@
-export const publicPagePaths = [
-    '/',
-    '/pricing',
-    '/docs',
-    '/docs/api',
-    '/docs/api/organizations',
-    '/docs/api/applications',
-    '/docs/api/self-hosted',
-    '/docs/sdk',
-    '/docs/sdk/environments',
-    '/docs/sdk/routes',
-    '/docs/sdk/storage',
-    '/docs/sdk/database',
-    '/docs/sdk/testing',
-    '/docs/sdk/building',
-    '/docs/sdk/pages',
-    '/docs/sdk/pages/longlink',
-    '/docs/sdk/pages/state',
-    '/docs/sdk/pages/query',
-    '/docs/sdk/pages/for',
-    '/docs/sdk/pages/action',
-    '/docs/sdk/pages/if',
-    '/docs/sdk/pages/i18n',
-    '/docs/sdk/pages/values',
-    '/docs/sdk/pages/count',
-    '/docs/sdk/pages/expressions',
-    '/docs/sdk/pages/bindings',
-    '/docs/sdk/pages/translations',
-    '/docs/sdk/pages/dynamic-pages',
-    '/docs/sdk/pages/page-files',
-    '/docs/sdk/pages/avatar',
-    '/docs/sdk/pages/badge',
-    '/docs/sdk/pages/banner',
-    '/docs/sdk/pages/button',
-    '/docs/sdk/pages/button-group',
-    '/docs/sdk/pages/card',
-    '/docs/sdk/pages/checkbox-input',
-    '/docs/sdk/pages/code',
-    '/docs/sdk/pages/dialog',
-    '/docs/sdk/pages/divider',
-    '/docs/sdk/pages/file-input',
-    '/docs/sdk/pages/form-layout',
-    '/docs/sdk/pages/grid',
-    '/docs/sdk/pages/heading',
-    '/docs/sdk/pages/icon',
-    '/docs/sdk/pages/link',
-    '/docs/sdk/pages/number-input',
-    '/docs/sdk/pages/radio-list',
-    '/docs/sdk/pages/radio-list-item',
-    '/docs/sdk/pages/selector',
-    '/docs/sdk/pages/selector-option',
-    '/docs/sdk/pages/side-nav',
-    '/docs/sdk/pages/side-nav-item',
-    '/docs/sdk/pages/slider',
-    '/docs/sdk/pages/stack',
-    '/docs/sdk/pages/switch',
-    '/docs/sdk/pages/tab',
-    '/docs/sdk/pages/tab-list',
-    '/docs/sdk/pages/table',
-    '/docs/sdk/pages/table-column',
-    '/docs/sdk/pages/text',
-    '/docs/sdk/pages/text-area',
-    '/docs/sdk/pages/text-input',
-    '/terms',
-    '/privacy',
-    '/impressum',
-] as const;
