+++ web/src/platform/routes/docs.tsx
@@ -1,14 +1,21 @@
-import type { MetaFunction } from 'react-router';
+import { useLocation, type MetaFunction } from 'react-router';
+import NotFound from '@/platform/NotFound';
+import DocsLayout from '@/platform/docs/layout';
 import { DOC_PAGES } from '@/platform/docs/catalog';
-import DocumentationRoutes from '@/platform/docs/routes';
-import { articleSeoPage, publicSeoMeta } from '@/lib/seo';
+import { noIndexMeta, publicSeoMeta } from '@/lib/seo';
 
 /** Returns metadata for the documentation article matched by the current URL. */
 export const meta: MetaFunction = ({ location }) => {
     const pathname = location.pathname.replace(/\/+$/, '') || '/';
     const page = DOC_PAGES.find((item) => item.path === pathname);
 
-    return page ? publicSeoMeta(articleSeoPage(page)) : [];
+    return page ? publicSeoMeta(page) : noIndexMeta('Not Found | LongLink');
 };
 
-export default DocumentationRoutes;
+/** Resolves and renders the documentation article matched by the current URL. */
+export default function DocumentationRoute() {
+    const pathname = useLocation().pathname.replace(/\/+$/, '') || '/';
+    const page = DOC_PAGES.find((item) => item.path === pathname);
+
+    return page ? <DocsLayout page={page} /> : <NotFound />;
+}
