+++ web/src/platform/routes/home.tsx
@@ -1,7 +1,8 @@
 import type { MetaFunction } from 'react-router';
 import Home from '@/platform/Home';
-import { homeSeoPage, publicSeoMeta } from '@/lib/seo';
+import { publicSeoMeta } from '@/lib/seo';
+import { homePage } from '@/platform/public';
 
-export const meta: MetaFunction = () => publicSeoMeta(homeSeoPage);
+export const meta: MetaFunction = () => publicSeoMeta(homePage);
 
 export default Home;
