+++ web/src/platform/routes/legal.tsx
@@ -1,14 +1,21 @@
-import type { MetaFunction } from 'react-router';
-import LegalRoutes from '@/platform/legal/routes';
+import { useLocation, type MetaFunction } from 'react-router';
+import NotFound from '@/platform/NotFound';
+import LegalLayout from '@/platform/legal/layout';
 import { LEGAL_PAGES } from '@/platform/legal/catalog';
-import { articleSeoPage, publicSeoMeta } from '@/lib/seo';
+import { noIndexMeta, publicSeoMeta } from '@/lib/seo';
 
 /** Returns metadata for the legal article matched by the current URL. */
 export const meta: MetaFunction = ({ location }) => {
     const pathname = location.pathname.replace(/\/+$/, '') || '/';
     const page = LEGAL_PAGES.find((item) => item.path === pathname);
 
-    return page ? publicSeoMeta(articleSeoPage(page)) : [];
+    return page ? publicSeoMeta(page) : noIndexMeta('Not Found | LongLink');
 };
 
-export default LegalRoutes;
+/** Resolves and renders the legal article matched by the current URL. */
+export default function LegalRoute() {
+    const pathname = useLocation().pathname.replace(/\/+$/, '') || '/';
+    const page = LEGAL_PAGES.find((item) => item.path === pathname);
+
+    return page ? <LegalLayout page={page} /> : <NotFound />;
+}
