+++ web/src/platform/routes/pricing.tsx
@@ -1,7 +1,8 @@
 import type { MetaFunction } from 'react-router';
 import Pricing from '@/platform/Pricing';
-import { pricingSeoPage, publicSeoMeta } from '@/lib/seo';
+import { publicSeoMeta } from '@/lib/seo';
+import { pricingPage } from '@/platform/public';
 
-export const meta: MetaFunction = () => publicSeoMeta(pricingSeoPage);
+export const meta: MetaFunction = () => publicSeoMeta(pricingPage);
 
 export default Pricing;
